% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transition_disruption_metric.R
\name{calculate_transition_disruption_metric}
\alias{calculate_transition_disruption_metric}
\title{Calculate the transition disruption metric, based on the IPR scenario}
\usage{
calculate_transition_disruption_metric(data, start_year, ...)
}
\arguments{
\item{data}{A dataset like the \verb{*_results_portfolio.rda} outputs of PACTA.}

\item{start_year}{The start year for which the TDM value should be calculated.}

\item{...}{Variables to group by.}
}
\value{
A tibble with the columns specified in the \code{...} input, as well as:
\code{tdm_tech}, the technology level transition disruption metric and \code{tdm_sec},
the sector level transition disruption metric.
}
\description{
Calculate the transition disruption metric, based on the IPR scenario
}
\examples{
pacta_results <- tibble::tibble(
  investor_name = rep("some_investor", 3),
  portfolio_name = rep("some_portfolio", 3),
  allocation = rep("portfolio_weight", 3),
  equity_market = rep("GlobalMarket", 3),
  scenario_geography = rep("Global", 3),
  ald_sector = rep("Power", 3),
  technology = rep("RenewablesCap", 3),
  year = c(2020, 2025, 2030),
  plan_alloc_wt_tech_prod = rep(1, 3),
  scen_alloc_wt_tech_prod = rep(1, 3),
  plan_carsten = rep(0.5, 3)
)
pacta_results

calculate_transition_disruption_metric(pacta_results, start_year = 2020)
}
