% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if a named object contains expected names}
\usage{
check_crucial_names(x, expected_names)
}
\arguments{
\item{x}{A named object.}

\item{expected_names}{String; expected names of \code{x}.}
}
\value{
Invisible \code{x}, or an error with informative message.
}
\description{
It is based on:
\url{https://www.rdocumentation.org/packages/fgeo.tool/versions/1.2.5/topics/check_crucial_names}.
}
\examples{
x <- c(a = 1)
check_crucial_names(x, "a")
try(check_crucial_names(x, "bad"))

data <- data.frame(a = 1)
check_crucial_names(data, "a")
try(check_crucial_names(data, "bad"))

# Applications for the error class "missing_names" ---------------------

tryCatch(
  check_crucial_names(x, "bad"),
  error = function(e) class(e)
)

# Wrapping in try() to allow running examples with no failure
try(
  # What's interesting is this
  tryCatch(
    check_crucial_names(x, "bad"),
    missing_names = function(e) {
      stop(
        "A different error message",
        call. = FALSE
      )
    }
  )
)

testthat::expect_error(
  check_crucial_names(x, "bad"),
  class = "missing_names"
)
}
\concept{miscellaneous utility functions}
