% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_error_checks.R
\name{check_grouped_portfolio_years}
\alias{check_grouped_portfolio_years}
\title{Check if a portfolio has a single holdings date}
\usage{
check_grouped_portfolio_years(port_holdings_date)
}
\arguments{
\item{port_holdings_date}{A character vector.}
}
\value{
Invisible \code{port_holdings_date}, or terminates the R Session with
\code{status = 0L}
}
\description{
This function exists to check if a portfolio has a single holdings
date, or multiple. If there is a single date, then everything is good.
but on the platform, if a user groups multiple  portfolios, and some of
these have different holding dates, then the parameters file will
contain a yaml array with the distinct dates from each grouped
portfolio.
}
\details{
Examples:

simple portfolio
holdings_date: 2020Q4

grouped portfolio with all same date:
holdings_date: 2020Q4

grouped portfolio with different dates:
holdings_date: \link{2019Q4, 2020Q4}

\code{config} will read this in as a character vector, so all we need to do
here is check for length > 1
}
\examples{
# No errors
check_grouped_portfolio_years("2019Q4")
check_grouped_portfolio_years(c("2019Q4"))
check_grouped_portfolio_years(c("2019Q4", "2019Q4"))
# produces error
\dontrun{check_grouped_portfolio_years(c("2019Q4", "2020Q4"))}
}
