% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_column_names.R
\name{clean_quo}
\alias{clean_quo}
\alias{clean_quo.quosure}
\alias{clean_quo.quosures}
\title{Clean quosure expressions}
\usage{
clean_quo(quo)

\method{clean_quo}{quosure}(quo)

\method{clean_quo}{quosures}(quo)
}
\arguments{
\item{quo}{An object of class quosure or of class quosures.}
}
\value{
A quosure or quosures.
}
\description{
Clean quosure expressions
}
\examples{
clean_quo(rlang::quo(x.x))

clean_quo(rlang::quos(x.x, y.y))
}
\seealso{
Other functions to deal with unclean names: 
\code{\link{clean_column_names}()},
\code{\link{quo_chr}()},
\code{\link{unclean_column_names}()}
}
\concept{functions to deal with unclean names}
