% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_delimiter.R
\name{guess_delimiter}
\alias{guess_delimiter}
\title{Guess the delimiter of a delimited file for a vector of filenames or
filepaths}
\usage{
guess_delimiter(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A character vector the same length as \code{filepaths}.
}
\description{
This function will guess the delimiter of a delimited file for a vector of
filenames or filepaths and return the delimiter as a string. It primarily
uses \code{vroom:::guess_delim()} to guess, and defaults to "," if that fails. If
the file is inaccessible or binary, it will return \code{NA} for that element.
}
