% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_file_encoding.R
\name{guess_file_encoding}
\alias{guess_file_encoding}
\title{Guess the file encoding for a vector of filenames or filepaths}
\usage{
guess_file_encoding(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A character vector the same length as \code{filepaths}.
}
\description{
This function will guess the file encoding of a vector of filenames or
filepaths and return the file encoding as a string. It primarily uses
\code{readr::guess_encoding()} with the \code{n_max = -1L} option so that all lines
are used to guess the encoding. If a file is inaccessible or if it is a
binary file, it will return \code{NA} for that element.
}
