% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary_file.R
\name{is_binary_file}
\alias{is_binary_file}
\title{Guess if a file is binary for a vector of filenames or filepaths}
\usage{
is_binary_file(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A logical vector the same length as \code{filepaths}.
}
\description{
This function will guess if a file is binary for a vector of filenames or
filepaths and return \code{TRUE} for each if so. It uses a vectorized version of
\code{file -b --mime-encoding} through \code{system2()} to guess. If the file is
inaccessible, it will return \code{NA} for that element.
}
