% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_file_accessible.R
\name{is_file_accessible}
\alias{is_file_accessible}
\title{Validate a vector of filenames or filepaths}
\usage{
is_file_accessible(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A logical vector the same length as \code{filepaths}.
}
\description{
This function validates that a vector of filenames or filepaths are
accessible files that: are a file, exist, have read access, and are not
empty. Dropbox files that are visible but not downloaded locally will be
empty files, and they will not pass this validation.
}
