% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_readable_file.R
\name{is_readable_file}
\alias{is_readable_file}
\title{Validate read access to files in a vector of filenames or filepaths}
\usage{
is_readable_file(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A logical vector the same length as \code{filepaths}.
}
\description{
This function validates that files in a vector of filenames or filepaths are
readable files and returns \code{TRUE} or \code{FALSE} for each one.
}
