% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_text_file.R
\name{is_text_file}
\alias{is_text_file}
\title{Guess if a file is a text file for a vector of filenames or filepaths}
\usage{
is_text_file(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A logical vector the same length as \code{filepaths}.
}
\description{
This function will guess if a file is a text file for a vector of filenames
or filepaths and return \code{TRUE} or \code{FALSE} for each. It primarily uses the
system's \code{file -b --mime-type} to guess the content type, but if \code{file} is
not found on the system's path (e.g. on Windows without Rtools installed)
and/or if the \code{find} path is not executable it will fall back on using
\code{wand::get_content_type()}, which is R-only and cross platform compatible
but less robust.
}
