% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_currency_code.R
\name{is_valid_currency_code}
\alias{is_valid_currency_code}
\title{Validate a vector of currency codes}
\usage{
is_valid_currency_code(currency_codes)
}
\arguments{
\item{currency_codes}{A character vector}
}
\value{
A logical vector the same length as \code{currency_codes}.
}
\description{
This function validates that a vector of currency codes are valid currency
codes that exist in the ISO 4217 alpha code specification.
}
