% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_cusip.R
\name{is_valid_cusip}
\alias{is_valid_cusip}
\title{Validate a vector of CUSIPs}
\usage{
is_valid_cusip(cusips)
}
\arguments{
\item{cusips}{A character or numeric vector (automatically collapses a single
column data.frame to a vector)}
}
\value{
A logical vector the same length as \code{cusips}.
}
\description{
This function validates that a vector of CUSIPs are valid CUSIP codes.
}
