% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_column_names.R
\name{quo_chr}
\alias{quo_chr}
\alias{quo_chr.list}
\alias{quo_chr.quosure}
\title{Convert quosure expressions to character}
\usage{
quo_chr(quo)

\method{quo_chr}{list}(quo)

\method{quo_chr}{quosure}(quo)
}
\arguments{
\item{quo}{A quosure or list with one quosure per list element.}
}
\value{
A character vector.
}
\description{
Convert quosure expressions to character
}
\examples{
quo <- rlang::quo(x)
quo_chr(quo)

# This is a list of quosure as crated via `...`, not a quosures
quo_chr(list(quo, quo))
}
\seealso{
Other functions to deal with unclean names: 
\code{\link{clean_column_names}()},
\code{\link{clean_quo}()},
\code{\link{unclean_column_names}()}
}
\concept{functions to deal with unclean names}
