% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_or_rda.R
\name{rds_or_rda}
\alias{rds_or_rda}
\title{Determines if a file is a .rda or .rds file}
\usage{
rds_or_rda(filepaths)
}
\arguments{
\item{filepaths}{A character vector}
}
\value{
A character vector the same length as \code{filepaths}.
}
\description{
This function will determine if a file is a \code{.rds} or \code{.rda}/\code{.RData} file
and return \code{"rds"} or \code{"rda"} respectively, or return \code{NA} if it is neither
or if the file is inaccessible. It uses the magic bits of the file to make a
determination, so it is not dependent on the file name extension, and it
automatically adapts to whether the file was saved as binary or ascii and if
it is compressed or not.
}
