% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_portfolio_csv.R
\name{read_portfolio_csv}
\alias{read_portfolio_csv}
\title{Read in portfolio CSV/s, working around a number of non-standard issues}
\usage{
read_portfolio_csv(filepaths, combine = TRUE)
}
\arguments{
\item{filepaths}{A character vector or single column data frame (strings
should be valid file paths to CSV files or a directory that contains CSV
files)}

\item{combine}{A single element logical (default \code{TRUE})}
}
\value{
If \code{combine} is \code{TRUE}, returns a tbl_df with all of the readable
data from the portfolio CSVs combined. If \code{combine} is \code{FALSE}, returns a
list of tbl_dfs, one for each readable portfolio CSV.
}
\description{
This function will read in one more portfolio CSVs. It works around a number
of common issues, like alternate column names, alternate delimiter,
alternate decimal and grouping marks, file encodings besides ASCII or UTF-8,
etc.
}
