% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_file.R
\name{write_csv_file}
\alias{write_csv_file}
\title{A version of readr::write_csv() where \code{file} falls back to \code{path} if needed}
\usage{
write_csv_file(x, file, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{File or connection to write to. In readr < 1.4.0 the name of this
argument is instead \code{path}. This wrapper helps use \code{file} regardless the
version of readr.}

\item{...}{
  Arguments passed on to \code{\link[readr:write_delim]{readr::write_csv}}
  \describe{
    \item{\code{na}}{String used for missing values. Defaults to NA. Missing values
will never be quoted; strings with the same value as \code{na} will
always be quoted.}
    \item{\code{append}}{If \code{FALSE}, will overwrite existing file. If \code{TRUE},
will append to existing file. In both cases, if the file does not exist a new
file is created.}
    \item{\code{col_names}}{If \code{FALSE}, column names will not be included at the top of the file. If \code{TRUE},
column names will be included. If not specified, \code{col_names} will take the opposite value given to \code{append}.}
    \item{\code{quote_escape}}{The type of escaping to use for quoted values, one of
\code{"double"}, \code{"backslash"} or \code{"none"}. You can also use \code{FALSE}, which is
equivalent to "none". The default is \code{"double"}, which is expected format for Excel.}
    \item{\code{eol}}{The end of line character to use. Most commonly either \code{"\n"} for
Unix style newlines, or \code{"\r\n"} for Windows style newlines.}
    \item{\code{path}}{\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}}
  }}
}
\description{
A version of readr::write_csv() where \code{file} falls back to \code{path} if needed
}
