% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tdm.R
\name{calculate_tdm}
\alias{calculate_tdm}
\title{Calculate and add the transition disruption metric to a dataset}
\usage{
calculate_tdm(data, t0, t1 = 5, t2 = 10, additional_groups = NULL)
}
\arguments{
\item{data}{A dataset, like the \verb{*_results_portfolio.rda} outputs of PACTA.}

\item{t0}{The start year, from which the TDM value should be calculated.}

\item{t1}{The number of years into the future for which there is production
data. Default is 5 years.}

\item{t2}{The number of years into the future against which you want to
calculate disruption. Default is 10 years.}

\item{additional_groups}{Character vector. The names of columns to group by,
in addition to these ones (which are always used):

technology, ald_sector}
}
\value{
A tibble with the columns specified in the \code{...} input, as well as:
\code{tdm_tech}, the technology level transition disruption metric and
\code{tdm_sec}, the sector level transition disruption metric.
}
\description{
This function takes a typical PACTA output, and uses it to calculate the
transition disruption metric.
}
\examples{
library(tibble)

pacta_results <- tibble(
  investor_name = rep("some_investor", 3),
  portfolio_name = rep("some_portfolio", 3),
  allocation = rep("portfolio_weight", 3),
  equity_market = rep("GlobalMarket", 3),
  scenario_geography = rep("Global", 3),
  ald_sector = rep("Power", 3),
  technology = rep("RenewablesCap", 3),
  year = c(2020, 2025, 2030),
  plan_alloc_wt_tech_prod = rep(1, 3),
  scen_alloc_wt_tech_prod = 1:3,
  plan_carsten = rep(0.5, 3)
)
pacta_results

calculate_tdm(pacta_results, t0 = 2020)
}
