% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transition_disruption_metric.R
\name{calculate_transition_disruption_metric}
\alias{calculate_transition_disruption_metric}
\title{Calculate the transition disruption metric, based on the IPR scenario}
\usage{
calculate_transition_disruption_metric(data, start_year, ...)
}
\arguments{
\item{data}{A dataset like the \verb{*_results_portfolio.rda} outputs of PACTA.}

\item{start_year}{The start year for which the TDM value should be calculated.}

\item{...}{Variables to group by.}
}
\value{
A tibble with the columns specified in the \code{...} input, as well as:
\code{tdm_tech}, the technology level transition disruption metric and \code{tdm_sec},
the sector level transition disruption metric.
}
\description{
Calculate the transition disruption metric, based on the IPR scenario
}
