% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_file.R
\name{write_csv_file}
\alias{write_csv_file}
\title{A version of readr::write_csv() where \code{file} falls back to \code{path} if needed}
\usage{
write_csv_file(x, file, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{File or connection to write to. In readr < 1.4.0 the name of this
argument is instead \code{path}. This wrapper helps use \code{file} regardless the
version of readr.}

\item{...}{Other arguments passed to readr::write_csv.}
}
\description{
A version of readr::write_csv() where \code{file} falls back to \code{path} if needed
}
